requirements.txt
asgiref==3.4.1
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.7
defusedxml==0.7.1
dj-database-url==0.5.0
Django==3.2.5
django-allauth==0.44.0
django-cors-headers==3.7.0
django-crispy-forms==1.12.0
django-heroku==0.3.1
django-restframework==0.0.1
djangorestframework==3.12.4
fpdf==1.7.2
gunicorn==20.1.0
idna==2.10
oauthlib==3.1.1
Pillow==8.3.0
psycopg2==2.9.1
psycopg2-binary==2.8.4
pycparser==2.20
PyJWT==2.1.0
python3-openid==3.2.0
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
sqlparse==0.4.1
typing-extensions==3.10.0.0
urllib3==1.26.6
whitenoise==5.2.0
xlrd==2.0.1
xlwt==1.3.0
